Sqlite database now stores only relative paths for diskitems in the minf representation (attributes name abnd _files) and in filenames table. This enables to copy a database to another location without having to update. Diskitem api is still the same, the paths are converted to absolute path when a diskitem is got from the database.
SQLDatabase now takes in parameters a directory instead of a list of directories. This feature was not used and prevent from storing relative paths in the database.
-> calls for SQLDatabase constructor have been modified.

The sqlite file that stores the database can now have a version : database-1.1.sqlite. The version used is fixed in neuroHierarchy and there is a mapping between database version and axon version.

When the sqlite database file doesn't exist, it isn't created automatically anymore. Indeed, updating databases can take a long time and should be done on user's demand. Now the user is warned that database needs to be update even for databases that missed sqlite file. If the user doesn't update the databases, he will have warning when he tries to insert or get diskitem from this database.

#566



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/axon/branches/3.2@35056 613412ce-1361-4b7a-93bf-a3c75af1bae4