Fixes #3874
Indeed, according to python documentation, the order in which the functions registered in atexit are called is not defined. So it is better to register only one function which calls the other in the correct order. 
I removed various atexit.register and replaced them by a cleanup function in neuro.py which is registered in atexit.
A cleanup function is also added in brainvisa.axon.processes.



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/axon/trunk@46120 613412ce-1361-4b7a-93bf-a3c75af1bae4