WIP on retreiving history/logs on distributed executions. Output databases involved in a process are listed. History info can be saved in all of those databases. Options are passed to the brainvisa commandline to record info in the correct history directories (python -m brainvisa.axon.runprocess --historyBook <path> [--historyBooko <otherpath>] ...). Workflows creation can handle these additional options, with additional FileTransfers to retreive history files from a distant processing resource.
It _almost_ works...
BUT: processes sometimes end up with an error "operation on closed file" when logging something while leaving BV. It is not reproducible. Re-running the workflow works.
SO: workflows handling history are still disabled until we find out the problem.
To enable it and try it, in workflow.py, just uncomment the 2 lines # self._handleHistoryBook( fileName )

#4570



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/axon/trunk@48119 613412ce-1361-4b7a-93bf-a3c75af1bae4