Fixes #4902
The problem was due to a modification for compatibity with PySide: replacing of a signal noDefault by noDefault(QString). It does not work with PyQt.
I have reverted this modification. For compatibility with PySide, we will have to find another solution like for example using the new style of signals and slots. But we have to be careful, this signal is used in several editors class.



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/axon/trunk@49044 613412ce-1361-4b7a-93bf-a3c75af1bae4