Fix using requireAttributes to require the absence of an attribute
We can now use this idiom:

    ReadDiskItem(..., requiredAttributes={'attr': ''})

to require that the selected DiskItem shall NOT have the 'attr' attribute. This
should not cause any conflicts because empty attribute values are not accepted.

Before this patch, this idiom worked correctly in the disk item browser window,
but upon clicking ok the DiskItem was rejected, because the diskItemFilter
function enforced the presence of all attributes in requiredAttributes.



git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/axon/branches/bug_fix@83668 613412ce-1361-4b7a-93bf-a3c75af1bae4