Avoid using the GUI to show exceptions until it is ready
Exceptions thrown by processes that fail to load used to be displayed in a GUI
pop-up window. The problem is that until the Qt event loop is started, this
pop-up is frozen and stays empty. The logging of errors to the GUI also
incurred very long startup times when there are lots of errors (such as when
AIMS fails to load due to a Python version mismatch). As a result, BrainVISA
appears to be hanging and refusing to start up.

This commit changes the error display mechanism (showException) to print to
stdout during start-up, switching to graphical pop-ups only when the event loop
is actually started.

Errors are still logged to the BrainVISA error log, so they can be displayed in
the GUI when it is initialized.

This is part of the fix for https://github.com/brainvisa/axon/issues/20