completely changed how the edition process works
not it works like a viewer: the process ends as soon as objects are
loaded, and callbacks are attached to a small object which is
responsible for cleaning things up.
Closing the edit mode can be echieved either by un-checking the pen
icon, or by closing the anatomist window.
No nested event loop or polling thread is used any longer.