revert last change 10668ad1c52701a4d4aae0cdfe58ff82b050b1c4 because it has side effects
some processes link an input from another input. In a pipeline, if the
1st input is the output of a preceding process, then it behaves like a
write diskitem. In the inputèinput link, this write attribute
propagation is expected, so I set it back.
It's propbably not a good thing, but existing pipelines rely on this
behaviour.
Thus we sometimes end up with inputs which don't exist.